chore(replicache): add size-of-value perf test
And refactor benchmark workflows to use reusable workflows
